+++ TipBot/Logic/CommandsManager.cs
@@ -27,7 +27,7 @@ namespace TipBot.Logic
 
         /// <summary>Transfers <paramref name="amount"/> of money from <paramref name="sender"/> to <paramref name="userBeingTipped"/>.</summary>
         /// <exception cref="CommandExecutionException">Thrown when user supplied invalid input data.</exception>
-        public void TipUser(IUser sender, IUser userBeingTipped, double amount)
+        public void TipUser(IUser sender, IUser userBeingTipped, decimal amount)
         {
             this.logger.Trace("({0}:'{1}',{2}:'{3}',{4}:{5})", nameof(sender), sender.Id, nameof(userBeingTipped), userBeingTipped.Id, nameof(amount), amount);
 
@@ -84,13 +84,13 @@ namespace TipBot.Logic
             return depositAddress;
         }
 
-        public double GetUserBalance(IUser user)
+        public decimal GetUserBalance(IUser user)
         {
             this.logger.Trace("({0}:{1})", nameof(user), user.Id);
 
             DiscordUser discordUser = this.GetOrCreateUser(user);
 
-            double balance = discordUser.Balance;
+            decimal balance = discordUser.Balance;
 
             this.logger.Trace("(-):{0}", balance);
             return balance;
@@ -138,7 +138,7 @@ namespace TipBot.Logic
             return userExists;
         }
 
-        private void AssertBalanceIsSufficient(DiscordUser user, double balanceRequired)
+        private void AssertBalanceIsSufficient(DiscordUser user, decimal balanceRequired)
         {
             this.logger.Trace("({0}:'{1}',{2}:{3})", nameof(user), user, nameof(balanceRequired), balanceRequired);
 
@@ -162,7 +162,7 @@ namespace TipBot.Logic
             this.logger.Trace("(-)");
         }
 
-        private void AssertAmountPositive(double amount)
+        private void AssertAmountPositive(decimal amount)
         {
             this.logger.Trace("({0}:{1})", nameof(amount), amount);
